% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{JoiniButtonDatasets}
\alias{JoiniButtonDatasets}
\title{Join Multi-Temporal iButton Datasets}
\usage{
JoiniButtonDatasets(datalist)
}
\arguments{
\item{datalist}{List of objects created using ReadiButtonFolder}
}
\value{
zoo object
}
\description{
Join a list of objects created by ReadiButtonFolder into a single dataset as a zoo object
}
\examples{
    packagePath <- find.package("iButtonDataOrganizer", lib.loc=NULL, quiet = TRUE)
    pathToFolder1 <- paste(packagePath, "/extdata/iButtonData_JulOct2014_2hr", sep="")
    pathToFolder2 <- paste(packagePath, "/extdata/iButtonData_NovDec2014_2hr", sep="")

    #Note: Time can also be included as part of the date; Date/Time
    #formatted as "YYYY-MM-DD HH-MM-SS", based on 24 hour time
    JulOct2014 <- ReadiButtonFolder(path=pathToFolder1, rounding="2hrs",
     StartDate="2014-07-26", EndDate="2014-10-31")
    NovDec2014 <- ReadiButtonFolder(path=pathToFolder2, rounding="2hrs",
     StartDate="2014-11-03", EndDate="2014-12-12")
    iButtonDatasets <- list(JulOct2014, NovDec2014)

    CompilediButtonData <- JoiniButtonDatasets(iButtonDatasets)
}
\author{
Mike Treglia \email{mtreglia@gmail.com}
}

