% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DailyStats.R
\name{temp.agg.daily}
\alias{temp.agg.daily}
\title{Calculate Temperature Statistics by Day}
\usage{
temp.agg.daily(dataname, stat)
}
\arguments{
\item{dataname}{Name of time-series data, generally a set of iButton data files imported through ReadiButtonFolder}

\item{stat}{Desired statistic to be calcualted as a daily aggregate. Can be any of the following: "var", "mean", "max", "min", "sd", or in format of: "function(x) var(x)"}
}
\description{
Calculate various statistics from iButton datasets, by day.
}
\examples{
    packagePath <- find.package("iButtonDataOrganizer", lib.loc=NULL, quiet = TRUE)
    pathToFolder <- paste(packagePath, "/extdata/iButtonData_JulOct2014_2hr", sep="")

    JulOct2014 <- ReadiButtonFolder(path=pathToFolder, rounding="2hrs",
     StartDate="2014-07-26", EndDate="2014-10-31")

    #Calculate the Daily Variance for each logger from which the data were imported
    JulOct2014.DailyVar <- temp.agg.daily(JulOct2014, var)
}
\author{
Mike Treglia \email{mtreglia@gmail.com}
}

